Properly fixed incremental builds this time.
Was using <fileset> instead of <path> to describe the output directory in the classpath,
and for some reason or another this was causing AspectJ weaving to sometimes not get into
the output classes. Fortunately the unit tests implicitly check that AOP has been done,
so those failed appropriately. It's running fine now locally, which means it'll probably
fail on Bamboo.