A selection of funs.
Created ConditionalInterceptor to avoid running OSIV interceptor for /static/*
Changed the JDBC connection extractor to the class used by Portal in an attempt to fix the BLOB bug.
Did an actual fix to the BLOB bug - removed cascade rules from Feedback to FileAttachment. I believe this cascade
rule is what's causing the FileAttachment to get flushed twice or something, resulting in some operation on a
Blob object that doesn't want to be touched any more.