HFC-48 Dept admins can add and remove module managers from the UI.
Behaviour of the app is still undefined for module managers. Permissions are being enforced but there's
no home view for them to view their modules from. Will likely adjust the department module list to allow
them to see only the modules that they are allowed to, and remove links that don't apply to them.