A few changes to a test.
Removed unnecessary type declarations on vals, some semicolons.

The close() calls should be in a try/finally block really.
It doesn't matter so much for a test, but have changed it to
illustrate the Closeables.closeThis method which lets you create
a closeable object and have the closing of it handled for you.