HFC-243 Add restricted submissions checkbox.
If the box is checked and you're not in the enrolled students for that assignment,
you currently get bounced to the page not found page, same as any other permission
denied error. For this page, we should probably still show a skeleton version of the
assignment's page, because it's likely that over time many students will arrive at
such a page without the permissions they should have, and we should help them as much
as possible. Possibly even a button they can click that will automatically email the
module managers the relevant information.