Add basic SBT project.
Start of an SBT build project, so that you can have the option to build/test using SBT.
It currently doesn't do AspectJ weaving so many tests will fail normally. There is an
sbt-aspectj plugin which I haven't got working yet. You may be able to run sbt with the
correct javaagent and get it working that way.