HFC-377 Remove Command aspect.
Command now has an abstract template method, work(), which is implemented by subclasses. Code that uses a command will continue to call apply(), and that method now does logging and stuff around the work.

Currently it's possible to call work() directly, bypassing the logging and stuff. We don't want this to be possible, so this is open to improvement.