Rename Command.work() to Command.applyInternal().
In the absence of any way of preventing this method from being called on its own
rather than via apply(), decided via Nick K's suggestion to just name it
accordingly. The name applyInternal suggests that it's an internal method
not to be called directly.

Apologies to anyone currently having fun merging this with their work.