JRebel config for new tabula layout.
Coursework is pointing under its modules at the target/classes where Eclipse is set up to compile.

JRebel supports dependent jars where the rebel.xml is present in the jar, which is why there's
a rebel.xml in the common module as well now. JRebel will monitor the common project and redeploy
that jar when changes happen.