TAB-429 Use tokenising analyser for departments in profile index
Also add sysadmin option to scope profile reindex by department.

NB. This reintroduces touchedDepartments for searching (TAB-280), but as an OR with the old affiliatedDepartments field, so that it'll have existing behaviour until reindex, then seamlessly add the extended search. After full reindex, we should be able to remove that extra TermQuery.