TAB-718 Add the capability to release a small group
Groups can be released to tutors, students, or both.
Users access the function via the "Notify" menu option. Once a group has
been released it cannot be re-released to the same group of users.
Releasing sends an email to the relevant users to inform them that they
have been allocated to a group.