Fix intermittently failing test
If hibernate triggers a flush midway through this test for any reason,
then the test would fail with an unsaved object exception.

This happened because non of the fixture data (assignment, feedback,
module, etc) was persistent, but the command was attempting to persist the
MarkerFeedback.
Fixed by replacing the StateService (which calls
SaveOrUpdate) with a mock, since it's behaviour isn't important for the
test