Fix intermittently failing test
If hibernate triggers a flush midway through this test for any
reason, then the test would fail with an unsaved object exception.

This happened because non of the fixture data (assignment, feedback,
module, etc) was persistent, but the command was attempting to
persist the MarkerFeedback.

Fixed by making it possible to construct a StateService with a session
provider other than Daoisms (which hard-codes the session to be provided
by hibernate). This allows mocking the call to session.saveOrUpdate,
whilst maintaining the rest of the service's behaviour.