Refactor Command for easier composition
Introduce a pair of new traits, ComposableCommand and CommandInternals.
CommandInternals is designed to hold just the applyInternal logic and
any state that the command needs, without all the associated services
and logic that Command requires. CommandInternals instances can then be
mixed in with ComposableCommand to make a fully-formed command object.

Permissions Checking is likewise abstracted out so it can be tested
independently from the applyInternal logic. Notification was already
like that.

Added a set of tests in ComposableCommandTest which hopefully illustrate
how it's meant to be used.