Prevent intermittent failures in MemberDaoTests
MemberDaoTests inserts data into an HSQL DB, and depending on the order
of the tests, may end up with tests seeing each other's data and getting
confused.

Made it pass by marking all tests as @DirtiesContext, to force creation
of a new HSQLDB instance for each test, but it would be better (and
faster) to make each test roll back cleanly on completion.