Test and sanitise the SettingsMap trait a bit.
There weren't any tests to check that incompatible types fell back to defaults.
Also got rid of the self type that didn't really benefit anyone and
just made the trait harder to mix in. The chainable methods will no longer return
the actual type of the object, but we never rely on that anyway. Arguably the
methods shouldn't return self, because that gives the suggestion that it's returning
a modified copy of the original, when in reality we're modifying the settings.