TAB-1868 Improve performance of viewing monitoring points by agent.
The performance of this screen has been substantially improved by refactoring the StudentRelationship class to link directly to both a StudentCourseDetails (via an SCJ code) for the student, and a Member for the agent (using the MemberStudentRelationship subclass).

There is still some code that doesn't work exactly as we'd expect, in particular because we don't have any ExternalStudentRelationships at the moment there are things that would almost certainly break if we did.

By controlling the relationship and expressing it directly in Hibernate, we can instruct Hibernate to do eager fetches of these associations in the DAO objects.