TAB-4044 Work in progress to rebuild the indexing structure of Tabula using ElasticSearch.
I had originally intended for it to be possible to run both the existing
Lucene-based querying and the new index in parallel, but that proved to be
too confusing and awkward with the different versions of Lucene that come in
from a newer Elasticsearch library, so I chose to bite the bullet and instead
just start from empty indexes.

This will run on a standby before it's made live, which means that we can do
a full reindex on the standby and when it goes live it'll all be there as expected.