TAB-4143 Don't run sysadmin-invoked imports on web
Replaces some sysadmin controllers that invoke commands directly with calls to schedule
a new trigger referencing the same job. This means that the sysadmin controller returns
immediately (which might be confusing, we might want to change that) but it means that
any rules that we set in the job are enforced, i.e. you can't break the concurrent execution
guard that we have on all jobs at the moment.