TAB-5299 Tune performance of marker information on CM2 homepage
This caches the result of the CM2 workflow progress service, as JSON, and reads it back to progress objects when it comes out of the cache. We currently cache this for 24 hours but it's invalidated whenever a Command is executed with the assignment ID in question as a property; I'm sure we'll run into situations where this isn't the case (but they should be rare). At the moment the CM2 workflow progress doesn't have any time-sensitive states for *marking-related* workflow stages - if and when we extend this to the admin homepage and elsewhere, we'll start showing progress that is time sensitive, for example if a student hasn't submitted and is late, at which point we'll probably have to put a custom cache expiry in so that it isn't cached across the deadline or across any extension deadlines.