TAB-5745 Remove SmallGroup from permissions parents of StudentMember
This implements the behaviour described in a 3-year old FIXME in StudentMember, noting that because a SmallGroup doesn't completely contain a StudentMember, it's incorrect for it to be a PermissionsParent of one and has strange behaviour (as well as performance implications). I can't see any way that the change made would make performance _worse_, but it won't necessarily make it better because every permissions lookup against an individual will still require listing all of the small groups that they are a member of, and checking that the current user isn't a tutor for any events in those groups.