TAB-5818 Parallelise email sending
This uses Futures as opposed to just using .par as that means we can control the execution context, using the one defined in uk.ac.warwick.tabula.helpers.Futures. This also ensures that a brand new (read-only) Hibernate Session is opened for each thread, but we manually mark the RecipientNotificationInfo as not read-only and manually flush, which ensures that our changes are saved as expected.