Fix Environment autowiring.
Think this is a bug in our spring-scala-glue, where it assumes that any bean name it gets from getBeansByType can then be passed in to getBeanDefinition(name), but that doesn't seem to be the case for the Environment instance even though it returns a name of 'environment'. In the meantime we can use EnvironmentAware to get it.