TAB-7267 Begin refactoring 🍝
Working on extracting the idea that certain entities in Tabula have a membership group that is defined by a query on the member table, plus members added or removed manually. Also members can be placed into groups within an entity, and if members are removed from the entity, they should be removed from any groups they're in.